Resource management

+++ src/ngn/fs.hpp
@@ -18,8 +18,8 @@ std::string contents(const std::string &fileName, bool throws = true);
 bool contents(const std::string &fileName, const std::string &contents, bool throws = true);
 std::string find(const std::string &fileName, bool throws = true);
 std::string ext(const std::string &fileName);
-size_t size(const std::string &fileName);
-size_t size(std::ifstream &in);
+int64_t size(const std::string &fileName);
+int64_t size(std::ifstream &in);
 
 } // fs
